LayoutTests/platform/chromium/fast/regex/cross-frame-callable-expected.txt

PASS re('a') threw exception TypeError: object is not a function.
PASS re.exec('a') is ['a']
PASS successfullyParsed is true

TEST COMPLETE
DONE
